+++ Code/src/edu/colostate/cs/cs414/banqi/model/Status.java
@@ -1,0 +1,6 @@
+package edu.colostate.cs.cs414.banqi.model;
+
+public enum Status {
+	ACTIVE, INACTIVE
+}
+
