+++ Code/src/edu/colostate/cs/cs414/banqi/model/Type.java
@@ -1,0 +1,6 @@
+package edu.colostate.cs.cs414.banqi.model;
+
+public enum Type {
+	GENERAL, ADVISOR, ELEPHANT, CHARIOT, HORSE, CANNON, SOLDIER
+}
+
