--- Code/src/model/Color.java
@@ -1,6 +1,0 @@
-package model;
-
-public enum Color {
-	RED, BLACK
-}
-
