--- Code/src/model/Status.java
@@ -1,6 +1,0 @@
-package model;
-
-public enum Status {
-	ACTIVE, INACTIVE
-}
-
