--- Code/src/model/Type.java
@@ -1,6 +1,0 @@
-package model;
-
-public enum Type {
-	GENERAL, ADVISOR, ELEPHANT, CHARIOT, HORSE, CANNON, SOLDIER
-}
-
