--- Code/src/model/User.java
@@ -1,28 +1,0 @@
-package model;
-
-import java.util.ArrayList;
-
-import controller.Controller;
-
-public class User {
-	
-	private String username, email, password;
-	
-	public User() {
-		//Empty constructor for test purposes
-	}
-	
-	public User(String username, String email, String password) {
-		this.username = username;
-		this.email = email;
-		this.password = password;
-	}
-	
-	public String getEmail() {
-		return email;
-	}
-	
-	public String getUsername() {
-		return username;
-	}
-}
